feat(canopy): make the client's HTTP transport pluggable
CanopyClient is now generic over a CanopyTransport: a trait that takes an
http::Request and answers with an http::Response. Everything above the HTTP
layer — the OpenAPI-generated endpoint methods and wire types, gzipping,
non-2xx mapping, JSON parsing — is transport-agnostic, so a caller that
reaches canopy some other way (a proxy that isn't a plain HTTP proxy, an
in-process handler, a stub in tests) implements the trait and keeps the whole
typed interface.

The parameter defaults to ReqwestTransport, which is the tailscale/mTLS
reqwest client moved out of client.rs unchanged, so existing callers keep
`CanopyClient`, `new`/`with_urls`, `is_tailscale`/`refresh`/`renew`, and the
raw-requests escape hatch exactly as they were.
