+++ crates/canopy/Cargo.toml
@@ -19,13 +19,16 @@ workspace = true
 
 [dependencies]
 algae-cli = { version = "1.1.9", path = "../algae-cli" }
+async-trait = "0.1.89"
 base64 = "0.22.1"
 blake3 = "1.8.5"
 bon = "3.9.3"
+bytes = "1.12.1"
 flate2 = "1.0.34"
 futures = { workspace = true, features = ["executor"] }
 getrandom = "0.4.2"
 hickory-resolver = "0.26.1"
+http = "1.4.2"
 jiff = { version = "0.2.28", features = ["serde"] }
 machine-uid = "0.6.0"
 miette = { workspace = true }
